Bring some sanity into the glyph positioning code
The implementation seems to assume that the size of the glyph bitmap
is always the same as the size of the glyph bounding box. In general
that is not true. That becomes painfully apparent when trying to use
some fonts without anti-aliasing (e.g. "Courier New"). The new logic
is based on That Fabulous (freetype) Manual, which clearly describes
how these kinds of things - glyph bitmap placement - should be done.